requirements.txt
# FastAPI and web framework dependencies
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.18
starlette==0.47.1

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jwt==4.1.0
# Pydantic for data validation
pydantic==2.11.7
email-validator==2.1.1
pydantic-settings==2.10.1
pydantic_core==2.33.2

# Data processing and utilities
numpy==2.3.1
requests==2.32.4
PyYAML==6.0.2

# JSON and data handling
jsonpatch==1.33
jsonpointer==3.0.0
orjson==3.10.18

# Streamlit for frontend
streamlit==1.41.1

# Async dependencies
aiohttp==3.12.14
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
anyio==4.9.0

# Additional utilities
click==8.2.1
certifi==2025.7.9
charset-normalizer==3.4.2
idna==3.10
urllib3==2.5.0
attrs==25.3.0
frozenlist==1.7.0
multidict==6.6.3
yarl==1.20.1

# Data science and analysis
marshmallow==3.26.1
dataclasses-json==0.6.7

# Misc utilities
packaging==24.2
sniffio==1.3.1
sortedcontainers==2.4.0
typing_extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
mypy_extensions==1.1.0

# LangSmith for monitoring
langsmith==0.4.5

# Additional dependencies
tenacity==9.1.2
websocket-client==1.8.0
zstandard==0.23.0
propcache==0.3.2

# Web scraping (if needed)
beautifulsoup4==4.13.4
lxml==6.0.0
soupsieve==2.7

# HTTP tools
requests-toolbelt==1.0.0
httpcore==1.0.9
httpx-sse==0.4.1
h11==0.16.0

# Distro info
distro==1.9.0


black